chore(app): rename app to navigation
