+++ src/app/components/components.js
@@ -1,6 +1,5 @@
 import angular from 'angular';
 
-import Navigation from './navigation';
 import LogoButton from './logo-button';
 import Products from './products';
 import Applications from './applications';
@@ -14,19 +13,18 @@ import Help from './help';
 import Search from './search';
 
 const componentsModule =
-          angular.module('components', [
-            Navigation,
-            LogoButton,
-            Products,
-            Applications,
-            BusinessPartner,
-            BusinessPartnerMenu,
-            Shortcuts,
-            User,
-            Users,
-            UserMenu,
-            Help,
-            Search
-          ]);
+  angular.module('components', [
+    LogoButton,
+    Products,
+    Applications,
+    BusinessPartner,
+    BusinessPartnerMenu,
+    Shortcuts,
+    User,
+    Users,
+    UserMenu,
+    Help,
+    Search
+  ]);
 
 export default componentsModule.name;
