--- src/app/components/navigation/navigation.spec.js
@@ -1,7 +1,0 @@
-import NavigationController from './navigation.controller';
-
-describe ('NavigationController', () => {
-  it('should exist', () => {
-    expect(NavigationController).not.toBe(undefined);
-  });
-});
