+++ src/app/navigation.service.js
@@ -1,4 +1,4 @@
-import * as Actions from '../../actions';
+import * as Actions from './actions';
 
 class NavigationService {
   constructor($ngRedux, $cookies, moduleConnectorService) {
